Move linuxmt/fdreg.h to arch/fdreg.h
